requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
Pillow>=8.0.0
PyPDF2>=2.0.0
typing-extensions>=4.0.0
pathlib>=1.0.1